refactored to prevent the need for special browser workarounds
